# specs/004-recording-module/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Aither Recording Module API
  version: "1.0.0"
  description: |
    Video and audio recording module for capturing WLAN webcam streams,
    managing recordings, controlling playback, and streaming files.
    All endpoints require Clerk authentication.
servers:
  - url: http://localhost:3000
    description: Local development
paths:
...
  /api/recording/playback/stop:
    post:
      operationId: pausePlayback
      summary: Stop (pause) playback
      security:
        - clerkAuth: []
      x-required-role: [admin, api-client]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaybackCommandRequest"
      responses:
        "200":
          description: Playback paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackResponse"
        "404":
          description: No active player session or recording not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      description: Clerk session token
  responses:
    Unauthenticated:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: UNAUTHENTICATED
    Forbidden:
      description: Insufficient role permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: FORBIDDEN
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
...
    PlaybackCommandRequest:
      type: object
      required: [recordingId]
      properties:
        recordingId:
          type: string
...
    PlaybackResponse:
      type: object
      required: [status, position]
      properties:
        status:
          type: string
          enum: [playing, paused]
        position:
          type: number
          description: Current playback position in seconds
...